+++ ts/common/system.ts
@@ -1,5 +1,5 @@
 //
-// Copyright 2014-21 Volker Sorge
+// Copyright 2014-26 Volker Sorge
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,549 +14,26 @@
 // limitations under the License.
 
 /**
- * @file Basic interface functionality for the Speech Rule Engine.
+ * @file The overall system API for the Speech Rule Engine.
  * @author volker.sorge@gmail.com (Volker Sorge)
  */
 
-import { AuditoryDescription } from '../audio/auditory_description.js';
 
-import { Engine, EnginePromise, SREError } from './engine.js';
-import { setupEngine as setup, engineSetup as engine } from './engine_setup.js';
-import * as EngineConst from './engine_const.js';
-import { KeyCode } from './event_util.js';
-import * as FileUtil from './file_util.js';
-import * as ProcessorFactory from './processor_factory.js';
-import { OptionsList, WorkerStructure } from './processor_factory.js';
 import { SystemExternal } from './system_external.js';
+import * as EngineConst from '../common/engine_const.js';
 import { Variables } from './variables.js';
-import { standardLoader } from '../speech_rules/math_map.js';
-import * as SpeechGeneratorUtil from '../speech_generator/speech_generator_util.js';
-import { RebuildStree } from '../walker/rebuild_stree.js';
-import * as DomUtil from './dom_util.js';
+import { setupEngine } from '../api/control.js';
 
-import { ClearspeakPreferences } from '../speech_rules/clearspeak_preferences.js';
+export * from '../api/control.js';
+export * from '../api/string.js';
+export * from '../api/misc.js';
+export * as worker from '../api/worker.js';
+export * as file from '../api/file.js';
 
 /**
  * Version number.
  */
 export const version: string = Variables.VERSION;
-
-/**
- *  Setup Methods functionality.
- */
-
-/**
- * Method to setup and initialize the speech rule engine. Currently the
- * feature parameter is ignored, however, this could be used to fine tune the
- * setup.
- *
- * @param feature An object describing some setup features.
- * @returns The promise that resolves once setup is complete.
- */
-export async function setupEngine(feature: {
-  [key: string]: boolean | string;
-}) {
-  return setup(feature);
-}
-
-/**
- * Query the engine setup.
- *
- * @returns Object vector with all engine feature
- *     values.
- */
-export function engineSetup(): { [key: string]: boolean | string } {
-  return engine();
-}
-
-/**
- * Reset the engine options setup.
- */
-export function resetEngine() {
-  Engine.getInstance().reset();
-}
-
-/**
- * @returns True if engine is ready, i.e., unicode file for the current
- *     locale has been loaded.
- */
-export async function engineReady(): Promise<any> {
-  return setupEngine({}).then(() => EnginePromise.getall());
-}
-
-/**
- * Export of the standard locale loader for use in client functions.
- */
-export const localeLoader = standardLoader;
-
-// Naming convention:
-// Input is either an XML expression as a string or from a file.
-// Output:
-//  toSpeech: Aural rendering string.
-//  toSemantic: XML of semantic tree.
-//  toJson: Json version of the semantic tree.
-//  toEnriched: Enriched MathML node.
-//  toDescription: List of auditory descriptions.
-// Output for the file version are strings.
-// TODO: (sorge) Need an async versions of these.
-/**
- * Main function to translate expressions into auditory descriptions.
- *
- * @param expr Processes a given XML expression for translation.
- * @returns The aural rendering of the expression.
- */
-export function toSpeech(expr: string): string {
-  return processString('speech', expr);
-}
-
-/**
- * Function to translate MathML string into Semantic Tree.
- *
- * @param expr Processes a given MathML expression for translation.
- * @returns The semantic tree as Xml.
- */
-export function toSemantic(expr: string): Node {
-  return processString('semantic', expr);
-}
-
-/**
- * Function to translate MathML string into JSON version of the Semantic Tree.
- *
- * @param expr Processes a given MathML expression for translation.
- * @returns The semantic tree as Json.
- */
-// TODO (TS): Define the correct JSONType somewhere.
-export function toJson(expr: string): any {
-  return processString('json', expr);
-}
-
-/**
- * Translate input expression into JSON version of the Semantic Tree that
- * serializes the content nodes.
- *
- * @param expr Processes a given MathML expression for translation.
- * @returns The semantic tree as Json.
- */
-export function toVis(expr: string): any {
-  return processString('vis', expr);
-}
-
-/**
- * Main function to translate expressions into auditory descriptions.
- *
- * @param expr Processes a given Xml expression for translation.
- * @returns The auditory descriptions.
- */
-export function toDescription(expr: string): AuditoryDescription[] {
-  return processString('description', expr);
-}
-
-/**
- * Function to translate MathML string into semantically enriched MathML.
- *
- * @param expr Processes a given MathML expression for translation.
- * @returns The enriched MathML node.
- */
-export function toEnriched(expr: string): Element {
-  return processString('enriched', expr);
-}
-
-/**
- * Translates a number into its word form for the current locale.
- *
- * @param expr The number
- * @returns The word form of the number.
- */
-export function number(expr: string): string {
-  return processString('number', expr);
-}
-
-/**
- * Translates a number into its word ordinal for the current locale.
- *
- * @param expr The number
- * @returns The word ordinal of the number.
- */
-export function ordinal(expr: string): string {
-  return processString('ordinal', expr);
-}
-
-/**
- * Translates a number into a numeric ordinal for the current locale.
- *
- * @param expr The number
- * @returns The numeric ordinal of the number.
- */
-export function numericOrdinal(expr: string): string {
-  return processString('numericOrdinal', expr);
-}
-
-/**
- * Translates a vulgar fraction into its word representation for the current
- * locale.
- *
- * @param expr The number with divisor slash.
- * @returns The vulgar fraction in words.
- */
-export function vulgar(expr: string): string {
-  return processString('vulgar', expr);
-}
-
-/**
- * Processes an input string with the given processor.
- *
- * @param processor The name of the processor to call.
- * @param input The input string.
- * @returns The computed data structure.
- * @template T
- */
-function processString<T>(processor: string, input: string): T {
-  return ProcessorFactory.process(processor, input);
-}
-
-/**
- * Namespace for file processing methods.
- */
-export const file: Record<string, (input: string, output?: string) => any> = {};
-
-/**
- * Reads an xml expression from a file and returns its aural rendering to a
- * file.
- *
- * @param input The input filename.
- * @param opt_output The output filename if one is given.
- * @returns Promise that resolves on completion of the file operations.
- */
-file.toSpeech = function (input: string, opt_output?: string) {
-  return processFile('speech', input, opt_output);
-};
-
-/**
- * Reads an xml expression from a file and returns the XML for the semantic
- * tree to a file.
- *
- * @param input The input filename.
- * @param opt_output The output filename if one is given.
- * @returns Promise that resolves on completion of the file operations.
- */
-file.toSemantic = function (input: string, opt_output?: string) {
-  return processFile('semantic', input, opt_output);
-};
-
-/**
- * Function to translate MathML string into JSON version of the Semantic Tree
- * to a file.
- *
- * @param input The input filename.
- * @param opt_output The output filename if one is given.
- * @returns Promise that resolves on completion of the file operations.
- */
-file.toJson = function (input: string, opt_output?: string) {
-  return processFile('json', input, opt_output);
-};
-
-/**
- * Main function to translate expressions into auditory descriptions
- * a file.
- *
- * @param input The input filename.
- * @param opt_output The output filename if one is given.
- * @returns Promise that resolves on completion of the file operations.
- */
-file.toDescription = function (input: string, opt_output?: string) {
-  return processFile('description', input, opt_output);
-};
-
-/**
- * Function to translate MathML string into semantically enriched MathML in a
- * file.
- *
- * @param input The input filename.
- * @param opt_output The output filename if one is given.
- * @returns Promise that resolves on completion of the file operations.
- */
-file.toEnriched = function (input: string, opt_output?: string) {
-  return processFile('enriched', input, opt_output);
-};
-
-/**
- * Reads an xml expression from a file, processes with the given function and
- * returns the result either to a file or to stdout.
- *
- * @param processor The name of the processor to call.
- * @param input The input filename.
- * @param opt_output The output filename if one is given.
- * @returns The promise for the file process to complete.
- */
-export function processFile(
-  processor: string,
-  input: string,
-  opt_output?: string
-) {
-  switch (Engine.getInstance().mode) {
-    case EngineConst.Mode.ASYNC:
-      return processFileAsync(processor, input, opt_output);
-    case EngineConst.Mode.SYNC:
-      return processFileSync(processor, input, opt_output);
-    default:
-      throw new SREError(
-        `Cannot process files in ${Engine.getInstance().mode} mode`
-      );
-  }
-}
-
-/**
- * Synchronously reads an xml expression from a file, processes with the given
- * function and returns the result either to a file or to stdout in synchronous
- * mode.
- *
- * @param processor The name of the processor.
- * @param input The input filename.
- * @param opt_output The output filename if one is given.
- * @returns The result that has been written to the file.
- */
-function processFileSync(
-  processor: string,
-  input: string,
-  opt_output?: string
-) {
-  const expr = inputFileSync_(input);
-  const result = ProcessorFactory.output(processor, expr);
-  if (opt_output) {
-    try {
-      SystemExternal.fs.writeFileSync(opt_output, result);
-    } catch (_err) {
-      throw new SREError('Can not write to file: ' + opt_output);
-    }
-  }
-  return result;
-}
-
-/**
- * Reads an xml expression from a file. Throws exception if file does not
- * exist.
- *
- * @param file The input filename.
- * @returns The input string read from file.
- */
-function inputFileSync_(file: string): string {
-  let expr;
-  try {
-    expr = SystemExternal.fs.readFileSync(file, { encoding: 'utf8' });
-  } catch (_err) {
-    throw new SREError('Can not open file: ' + file);
-  }
-  return expr;
-}
-
-/**
- * Reads an xml expression from a file, processes with the given function and
- * returns the result either to a file or to stdout in asynchronous mode.
- *
- * @param processor The name of the processor.
- * @param file The input filename.
- * @param output The output filename if one is given.
- * @returns The result of that is written to the file.
- */
-async function processFileAsync(
-  processor: string,
-  file: string,
-  output?: string
-) {
-  const expr = await SystemExternal.fs.promises.readFile(file, {
-    encoding: 'utf8'
-  });
-  const result = ProcessorFactory.output(processor, expr);
-  if (output) {
-    try {
-      SystemExternal.fs.promises.writeFile(output, result);
-    } catch (_err) {
-      throw new SREError('Can not write to file: ' + output);
-    }
-  }
-  return result;
-}
-
-// These are still considered experimental.
-/**
- * Walk a math expression provided by an external system.
- *
- * @param expr The string containing a MathML representation.
- * @returns The initial speech string for that expression.
- */
-export function walk(expr: string): string {
-  return ProcessorFactory.output('walker', expr);
-}
-
-/**
- * Moves in the math expression that is currently being walked.
- *
- * @param direction The direction of the move
- *     given either as string or keycode.
- * @returns The speech string generated by the walk. Null if a boundary
- *     is hit.
- */
-export function move(direction: KeyCode | string): string | null {
-  return ProcessorFactory.keypress('move', direction);
-}
-
-/**
- * A clean exit method, that ensures all file processes are completed.
- *
- * @param opt_value The exit value. Defaults to 0.
- */
-export function exit(opt_value?: number) {
-  const value = opt_value || 0;
-  EnginePromise.getall().then(() => process.exit(value));
-}
-
-/**
- * Function to translate expression into speech structure, with all speech
- * components for all nodes.
- *
- * @param expr Processes a given MathML expression for translation.
- * @returns The json structure containing all speech mappings.
- */
-export function toSpeechStructure(expr: string): string {
-  return processString('speechStructure', expr);
-}
-
-/**
- *  Web worker related API methods.
- */
-
-/**
- * Compute speech structure for the expression.
- *
- * @param expr The math expression.
- * @param options The list of options.
- * @returns The worker structure once the promise resolves.
- */
-export async function workerSpeech(
-  expr: string,
-  options: OptionsList
-): Promise<WorkerStructure> {
-  const mml = DomUtil.parseInput(expr);
-  const rebuilt = new RebuildStree(mml);
-  const styles = SpeechGeneratorUtil.toStyles(options);
-  options.domain2style = SpeechGeneratorUtil.fromStyles(styles);
-  return assembleWorkerStructure(mml, rebuilt.stree.xml(), options);
-}
-
-/**
- * Computes the speech for the next rule set.
- *
- * @param expr The math expression.
- * @param options The list of options.
- * @returns The worker structure once the promise resolves.
- */
-export async function workerNextRules(
-  expr: string,
-  options: OptionsList
-): Promise<WorkerStructure> {
-  // TODO: Don't do anything if no next rules!
-  const mml = DomUtil.parseInput(expr);
-  const rebuilt = new RebuildStree(mml);
-  const styles = SpeechGeneratorUtil.toStyles(options);
-  options = SpeechGeneratorUtil.nextRules(options, styles);
-  options.domain2style = SpeechGeneratorUtil.fromStyles(styles);
-  return assembleWorkerStructure(mml, rebuilt.stree.xml(), options);
-}
-
-/**
- * Computes the speech for the next style wrt to a particular node.
- *
- * @param expr The math expression.
- * @param options The list of options.
- * @param id Semantic id of the focused node.
- * @returns The worker structure once the promise resolves.
- */
-export async function workerNextStyle(
-  expr: string,
-  options: OptionsList,
-  id: string
-): Promise<WorkerStructure> {
-  // TODO: Don't do anything if no next style!
-  const mml = DomUtil.parseInput(expr);
-  const rebuilt = new RebuildStree(mml);
-  const styles = SpeechGeneratorUtil.toStyles(options);
-  options.style = SpeechGeneratorUtil.nextStyle(rebuilt.nodeDict[id], options);
-  styles[options.domain] = options.style;
-  options.domain2style = SpeechGeneratorUtil.fromStyles(styles);
-  return assembleWorkerStructure(mml, rebuilt.stree.xml(), options);
-}
-
-/**
- * Compute clearspeak preferences for a locale.
- *
- * @param options The options containing the locale setting.
- * @returns The worker structure once the promise resolves.
- */
-export async function workerLocalePreferences(
-  options: OptionsList
-): Promise<WorkerStructure> {
-  return ClearspeakPreferences.getLocalePreferences()[options.locale];
-}
-
-/**
- * Compute clearspeak preference category for a node.
- *
- * @param expr The math expression.
- * @param id The semantic id of a node in the expression to compute the category for.
- * @returns The worker structure once the promise resolves.
- */
-export async function workerRelevantPreferences(
-  expr: string,
-  id: string
-): Promise<string> {
-  const mml = DomUtil.parseInput(expr);
-  const rebuilt = new RebuildStree(mml);
-  const query =
-    rebuilt.stree.root.querySelectorAll((x) => x.id.toString() === id)[0] ??
-    rebuilt.stree.root;
-  return ClearspeakPreferences.relevantPreferences(query);
-}
-
-/**
- * Computes the structure returnable to the worker, containing all necessary
- * speech content to be attached.
- *
- * @param mml The math expression.
- * @param sxml The element.
- * @param options The list of options.
- * @returns The worker structure once the promise resolves.
- */
-async function assembleWorkerStructure(
-  mml: Element,
-  sxml: Element,
-  options: OptionsList
-): Promise<WorkerStructure> {
-  await setupEngine(options);
-  Engine.getInstance().options.automark = true;
-  const json: WorkerStructure = {};
-  ProcessorFactory.assembleSpeechStructure(json, mml, sxml, options);
-  if ((options as any).enableBraille === false) {
-    return json;
-  }
-  await setupEngine({
-    modality: 'braille',
-    locale: options.braille,
-    domain: 'default',
-    style: 'default'
-  });
-  const root = (sxml.childNodes[0] as Element)?.getAttribute('id');
-  json.braille = SpeechGeneratorUtil.computeBrailleStructure(sxml);
-  json.braillelabel = json.braille[root]['braille-none'];
-  return json;
-}
-
-/**
- * Returns the default locale path, depending on the mode of operation.
- *
- * @param locale The locale iso.
- * @param ext An optional file extension. Defaults to json.
- */
-export const localePath = FileUtil.localePath;
 
 if (SystemExternal.documentSupported || SystemExternal.webworker) {
   setupEngine({ mode: EngineConst.Mode.HTTP }).then(() => setupEngine({}));
